Добавление функционала для управления пользователями: реализованы страницы для добавления, редактирования и удаления пользователей, а также интеграция с API для получения и обновления данных пользователей. Обновлены модели и роутеры для поддержки новых функций.

+++ brain-api/models/Scene.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Text
+from sqlalchemy import Integer, String, Text, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.dialects.postgresql import JSON
 from typing import Optional
@@ -10,14 +10,14 @@ from database.annotated_types import uuidPK, createAT, updateAT
 class Scene(Base, AsyncCRUDMixin):
     __tablename__ = "scenes"
 
-    user_id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False,
+    user_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False,
                                          index=True, primary_key=True
                                          )
 
     scene: Mapped[str] = mapped_column(String, nullable=False)
     scene_path: Mapped[str] = mapped_column(String, nullable=False)
     page: Mapped[str] = mapped_column(String, nullable=False)
-    message_id: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    message_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
     
     data: Mapped[dict] = mapped_column(JSON, nullable=False, default={})
